Create demo files for building books with R bookdown package.

Technically, bookdownplus is a collection and selector of R bookdown templates. bookdownplus helps you write academic journal articles, guitar books, chemical equations, mails, calendars, and diaries. R bookdownplus extends the features of bookdown, and simplifies the procedure. Users only have to choose a template, clarify the book title and author name, and then focus on writing the text. No need to struggle in YAML and LaTeX.

Before starting, you have to install bookdown package, and other software and packages (i.e. 'Pandoc', 'LaTeX', rmarkdown, rticle, knitr, etc.) which bookdown depends on. See the official manual of bookdown for details.

Run bookdownplus() and specify the template with template argument, You will get some files (e.g. index.Rmd, body.Rmd) and folders in your working directory. Although there are many other arguments for bookdownplus(), you can simply ignore them if you use bookdownplus package for the first time.

Build it with bookdown. Your will get a book file named *.pdf in _book/ folder.

Write your own text in index.Rmd and body.Rmd, and build your own lovely book.

bookdownplus(template = "copernicus", more_output = NULL, title = "title",
    author = "author", render = TRUE, rproj = TRUE, output_name = NA, to = "./",
    new = TRUE)

Arguments

template
character. Choose a template for your book or doc. The default template is 'thesis_classic'. Run template() to see the available templates.

more_output
optional output formats besides pdf_book, which can be (run get_output() for hints):

NULL (default)

'word_document2'

'html_document2'

'epub_book'

'gitbook'

title
character. book title.

author
character. book author.

render
logical. whether to render automatically

rproj
logical. whether to created an .Rproj file automatically

output_name
chracter. the name of the output files. If NA (default), the template name will be used.

Value

demo files to build with bookdown

Examples

bookdownplus(render = FALSE)