Available templates.

Visit http://b.pzhao.org to see a gallery of the templates.

get_template()

Value

a dataframe displaying available templates.

Examples

get_template()

#>            name                            author contributor
#> 1       classic                         Peng Zhao   Peng Zhao
#> 2    copernicus                         Peng Zhao   Peng Zhao
#> 3           crc                         Yihui Xie   Peng Zhao
#> 4          demo                         Yihui Xie   Peng Zhao
#> 5       demo_zh                         Yihui Xie   Peng Zhao
#> 6       docsens               Are Edvardsen, SKDE   Peng Zhao
#> 7      elsevier                         Peng Zhao   Peng Zhao
#> 8       journal                         Peng Zhao   Peng Zhao
#> 9          mail                         Peng Zhao   Peng Zhao
#> 10         mdpi                         Peng Zhao   Peng Zhao
#> 11         mini                         Yihui Xie   Peng Zhao
#> 12       nonpar          Eduardo García-Portugués   Peng Zhao
#> 13       nte_zh                         Peng Zhao   Peng Zhao
#> 14     paper_zh                         Peng Zhao   Peng Zhao
#> 15       pku_zh                         Peng Zhao   Peng Zhao
#> 16         poem                         Peng Zhao   Peng Zhao
#> 17      rbasics Chester Ismay, Patrick C. Kennedy   Peng Zhao
#> 18    thesis_zh                         Peng Zhao   Peng Zhao
#> 19          ubt                         Peng Zhao   Peng Zhao
#> 20      ucas_zh                         Peng Zhao   Peng Zhao
#> 21     calendar                         Peng Zhao   Peng Zhao
#> 22       cchess                         Peng Zhao   Peng Zhao
#> 23    chemistry                         Peng Zhao   Peng Zhao
#> 24 chemistry_zh                         Peng Zhao   Peng Zhao
#> 25     cumtb_zh                    Xiangyun Huang   Peng Zhao
#> 26          dnd                         Peng Zhao   Peng Zhao
#> 27       guitar                         Peng Zhao   Peng Zhao
#> 28          igo                         Peng Zhao   Peng Zhao
#> 29        musix                         Peng Zhao   Peng Zhao
#> 30         skak                         Peng Zhao   Peng Zhao
#> 31           uw                      Eddie Kasner   Peng Zhao
#> 32       zju_zh                         Peng Zhao   Peng Zhao
#>                                                                                                                                                        intro
#> 1                                                                                                                                            Classic thesis.
#> 2                                        Write academic papers of Copernicus journals. Support both the two-columned pdf version and the discussion version.
#> 3                                                                                Writing a book of [Chapman & Hall](https://github.com/yihui/bookdown-crc). 
#> 4                                                                          [The original demo book for bookdown](https://github.com/rstudio/bookdown-demo). 
#> 5                                              [The original demo book for bookdown in Chinese](https://github.com/yihui/bookdown-chinese). (官方中文模板。)
#> 6                                                   A template from the book [Lokal lagring og bruk av sensitive data](https://bookdown.org/areedv/docSens/)
#> 7                                                                                                                Write academic papers of Elsevier journals.
#> 8                                                                                                                      Laboratory journal or personal diary.
#> 9                                                                                            Write elegant traditional mails. Multiple themes are available.
#> 10                                                                                                     Write academic papers in the layout of MDPI journals.
#> 11                                                                              [A mini demo book for bookdown](https://github.com/yihui/bookdown-minimal). 
#> 12                              A book template from the book [A short course on nonparametric curve estimation](https://bookdown.org/egarpor/nonpar-eafit/)
#> 13                                                                                             A template from Nothing to envy, in Chinese. (中文散文模板。)
#> 14 Write academic manuscript in Chinese with bilingual titles, author lists, abstracts, and keywords. (中文学术论文模板。支持双语标题、作者、摘要、关键词。)
#> 15                                                                      Write a thesis of Peking University with R markdown syntax. (北京大学学位论文模板。)
#> 16                                                                                                                                              A poem book.
#> 17                                             A template from the book [Getting used to R, RStudio, and R Markdown](https://ismayc.github.io/rbasics-book/)
#> 18                                                                                       A simple template for thesis in Chinese. (简单的中文学位论文模板。)
#> 19                                                                                                                          Thesis of University of Bayreuth
#> 20                                                           Write a thesis of the University of Chinese Academy of Sciences. (中国科学院大学学位论文模板。)
#> 21                                                                                                                        Insert monthly or weekly calendar.
#> 22                                                                                                       Instert Chinese chess diagrams. (插入中国象棋棋谱.)
#> 23                                                                                              Write a book with chemical molecular formulae and equations.
#> 24                                                                                                Write a chemistry book in Chinese. (中文文档插入化学符号.)
#> 25                                                         Thesis of China University of Mining and Technology, Beijing. (中国矿业大学（北京）学位论文模板.)
#> 26                                                                                                                         I don't know what on earth it is.
#> 27                                                                                                                          Write a book with guitar chords.
#> 28                                                                                                                                  Insert Go game diagrams.
#> 29                                                                                                                      Typesetting music and related stuff.
#> 30                                                                                                                                    Insert chess diagrams.
#> 31                                                                                                                        Thesis of University of Washington
#> 32                                                                                        Thesis of Zhejiang Uni., in Chinese. (据说是浙江大学学位论文模板.)
#>                                                                             download
#> 1     https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/classic.zip
#> 2  https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/copernicus.zip
#> 3         https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/crc.zip
#> 4        https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/demo.zip
#> 5     https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/demo_zh.zip
#> 6     https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/docsens.zip
#> 7    https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/elsevier.zip
#> 8     https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/journal.zip
#> 9        https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/mail.zip
#> 10       https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/mdpi.zip
#> 11       https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/mini.zip
#> 12     https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/nonpar.zip
#> 13     https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/nte_zh.zip
#> 14   https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/paper_zh.zip
#> 15     https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/pku_zh.zip
#> 16       https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/poem.zip
#> 17    https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/rbasics.zip
#> 18  https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/thesis_zh.zip
#> 19        https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/ubt.zip
#> 20    https://github.com/pzhaonet/bookdownplus/raw/master/inst/templates/ucas_zh.zip
#> 21      https://github.com/pzhaonet/bookdownplus/raw/master/upload/calendar/demo.zip
#> 22        https://github.com/pzhaonet/bookdownplus/raw/master/upload/cchess/demo.zip
#> 23     https://github.com/pzhaonet/bookdownplus/raw/master/upload/chemistry/demo.zip
#> 24  https://github.com/pzhaonet/bookdownplus/raw/master/upload/chemistry_zh/demo.zip
#> 25      https://github.com/pzhaonet/bookdownplus/raw/master/upload/cumtb_zh/demo.zip
#> 26           https://github.com/pzhaonet/bookdownplus/raw/master/upload/dnd/demo.zip
#> 27        https://github.com/pzhaonet/bookdownplus/raw/master/upload/guitar/demo.zip
#> 28           https://github.com/pzhaonet/bookdownplus/raw/master/upload/igo/demo.zip
#> 29         https://github.com/pzhaonet/bookdownplus/raw/master/upload/musix/demo.zip
#> 30          https://github.com/pzhaonet/bookdownplus/raw/master/upload/skak/demo.zip
#> 31            https://github.com/eddiekasner/example-dissertation/archive/master.zip
#> 32        https://github.com/pzhaonet/bookdownplus/raw/master/upload/zju_zh/demo.zip
#>    location
#> 1     local
#> 2     local
#> 3     local
#> 4     local
#> 5     local
#> 6     local
#> 7     local
#> 8     local
#> 9     local
#> 10    local
#> 11    local
#> 12    local
#> 13    local
#> 14    local
#> 15    local
#> 16    local
#> 17    local
#> 18    local
#> 19    local
#> 20    local
#> 21   remote
#> 22   remote
#> 23   remote
#> 24   remote
#> 25   remote
#> 26   remote
#> 27   remote
#> 28   remote
#> 29   remote
#> 30   remote
#> 31   remote
#> 32   remote